refactor(api): move scdl repositories in sub folder
